[Bronze V] Title: 이상한 기호, Time: 128 ms, Memory: 14384 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze V/15964. 이상한 기호/이상한 기호.java
@@ -7,10 +7,10 @@ public class Main {
     public static void main(String[] args) throws IOException {
         try (BufferedReader br = new BufferedReader(new InputStreamReader(System.in))) {
             StringTokenizer st = new StringTokenizer(br.readLine());
-            int a = Integer.parseInt(st.nextToken());
-            int b = Integer.parseInt(st.nextToken());
+            long a = Long.parseLong(st.nextToken());
+            long b = Long.parseLong(st.nextToken());
 
-            int result = (a+b) * (a-b);
+            long result = (a+b) * (a-b);
 
             System.out.println(result);
         }
